image for wisata

+++ App/goToba/goToba/src/main/java/com/example/goToba/service/implement/WisataServiceImpl.java
@@ -2,10 +2,13 @@ package com.example.goToba.service.implement;
 
 import com.example.goToba.model.SequenceWisata;
 import com.example.goToba.model.Wisata;
+import com.example.goToba.payload.helper.StockKeepingUnit;
+import com.example.goToba.payload.imagePath.ImagePath;
 import com.example.goToba.payload.request.WisataRequest;
 import com.example.goToba.redis.template.RedisKeys;
 import com.example.goToba.repository.SequenceWisataRepo;
 import com.example.goToba.repository.WisataRepo;
+import com.example.goToba.service.ImageService;
 import com.example.goToba.service.WisataService;
 import com.example.goToba.service.redisService.WisataRedisService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +18,7 @@ import reactor.core.publisher.Mono;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.Base64;
 
 /**
@@ -31,6 +35,9 @@ public class WisataServiceImpl implements WisataService {
 
     @Autowired
     WisataRedisService wisataRedisService;
+
+    @Autowired
+    ImageService imageService;
 
     @Override
     public Mono<Wisata> addWisata(WisataRequest wisataRequest) {
@@ -67,13 +74,22 @@ public class WisataServiceImpl implements WisataService {
                             wisataRequest.getName(),
                             wisataRequest.getTitle(),
                             wisataRequest.getDescription(),
-                            "/get/" + data.getKey() + "_000" + Integer.parseInt(data.getLast_seq()) + ".png",
+                            ImagePath.IMAGE_PATH_WISATA + ImagePath.IMAGE_CONNECTOR + data.getKey() + StockKeepingUnit.SKU_CONNECTOR + StockKeepingUnit.SKU_DATA_BEGINNING + Integer.parseInt(data.getLast_seq()) + ImagePath.IMAGE_EXTENSION,
+                            wisataRequest.getLongitude(),
+                            wisataRequest.getLatitude(),
                             wisataRequest.getAddress(),
                             wisataRequest.getCreatedBy(),
                             wisataRequest.getPrice(),
                             wisataRequest.getHoursOpen(),
                             "active"
                     );
+                    if (wisataRequest.getImage() != "") {
+                        try {
+                            imageService.addPicture(wisataRequest.getImage(), wisata.getSkuWisata(), ImagePath.IMAGE_PATH_WISATA);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
                     wisataRedisService.add(wisata);
                     return wisataRepo.save(wisata);
                 });
@@ -105,14 +121,18 @@ public class WisataServiceImpl implements WisataService {
                             wisataRequest.getTitle(),
                             wisataRequest.getDescription(),
                             wisataRequest.getImage(),
+                            wisataRequest.getLongitude(),
+                            wisataRequest.getLatitude(),
                             wisataRequest.getAddress(),
                             wisataRequest.getCreatedBy(),
                             wisataRequest.getPrice(),
                             wisataRequest.getHoursOpen(),
                             data.getStatus()
                     );
-                    wisataRedisService.deleteByKey(sku);
-                    wisataRedisService.add(wisata);
+                    if (wisataRedisService.hasKey(sku)) {
+                        wisataRedisService.deleteByKey(sku);
+                        wisataRedisService.add(wisata);
+                    }
                     return wisataRepo.save(wisata);
                 });
     }
@@ -130,6 +150,8 @@ public class WisataServiceImpl implements WisataService {
                             data.getTitle(),
                             data.getDescription(),
                             data.getImage(),
+                            data.getLongitude(),
+                            data.getLatitude(),
                             data.getAddress(),
                             data.getCreatedBy(),
                             data.getPrice(),
@@ -142,21 +164,6 @@ public class WisataServiceImpl implements WisataService {
 
     @Override
     public Mono<Wisata> findBySku(String sku) {
-        return wisataRepo.findFirstBySkuWisata(sku)
-                .flatMap(req -> {
-                    Wisata wisata = new Wisata(
-                            sku,
-                            req.getName(),
-                            req.getTitle(),
-                            req.getDescription(),
-                            req.getImage(),
-                            req.getAddress(),
-                            req.getCreatedBy(),
-                            req.getPrice(),
-                            req.getHoursOpen(),
-                            "active"
-                    );
-                    return wisataRepo.findFirstBySkuWisata(sku);
-                });
+        return wisataRepo.findFirstBySkuWisata(sku);
     }
 }
